+++ packages/format-library/src/link/index.js
@@ -19,10 +19,11 @@ import { RichTextToolbarButton, RichTextShortcut } from '@wordpress/block-editor
 import InlineLinkUI from './inline';
 
 const name = 'core/link';
+const title = __( 'Link' );
 
 export const link = {
 	name,
-	title: __( 'Link' ),
+	title,
 	tagName: 'a',
 	className: null,
 	attributes: {
@@ -102,7 +103,7 @@ export const link = {
 					{ ! isActive && <RichTextToolbarButton
 						name="link"
 						icon="admin-links"
-						title={ __( 'Link' ) }
+						title={ title }
 						onClick={ this.addLink }
 						isActive={ isActive }
 						shortcutType="primary"