+++ packages/docgen/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Master
+
+### Internal
+
+- Remove unneccessary argument from an instance of `Array#pop`.
+
 ## 1.0.0 (2019-03-06)
 
 - Initial release